refactor: create TOTP module and insert in batch the recovery codes
